# openapi.yaml
openapi: 3.0.3
info:
  title: MathLearning API (partial)
  version: "1.0.0"
  description: >-
    Partial OpenAPI 3.0 spec generated from the project's endpoint mappings
    (extracted 2026-05-04). This is not exhaustive — verify types and add
    missing paths/schemas as needed.
  contact:
    name: MathLearning Dev
    email: dev@mathlearning.example.com
servers:
  - url: http://localhost:5180
    description: Development server (local)

# Require Bearer JWT by default; anonymous endpoints override with security: []
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SyncRequestDto:
      type: object
      additionalProperties: true
    RegisterSyncDeviceRequest:
      type: object
      properties:
        deviceId: { type: string }
        deviceName: { type: string }
        platform: { type: string }
      required: [deviceId]
...
paths:
...
  /api/sync:
    post:
      tags: [Sync]
      summary: Sync client payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequestDto'
      responses:
        '200':
          description: SyncResponseDto (implementation-defined)
  /api/devices/register:
    post:
      tags: [Sync]
      summary: Register sync device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterSyncDeviceRequest'
      responses:
        '200':
          description: Device registration info
  /api/offline/bundle:
    get:
      tags: [Sync]
      summary: Get offline bundle
      parameters:
        - name: subtopicId
          in: query
          schema: { type: integer }
        - name: questionCount
          in: query
          schema: { type: integer, default: 100 }
      responses:
        '200':
          description: Offline bundle (implementation-defined)
...
  /api/sync/checkpoint:
    get:
      tags: [Sync]
      summary: Get sync checkpoint for device
      parameters:
        - name: deviceId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Checkpoint info
          content:
            application/json:
              schema:
                type: object
...